+++ CMakeLists.txt
@@ -18,6 +18,7 @@ set(HEADER_FILES
   include/portable_endian.hpp)
 
 set(SOURCE_FILES
+  src/form.cpp
   src/variant.cpp
   src/msgpack.cpp
   src/json.cpp)
